requirements.txt
flask
requests
python-dotenv
fastapi 
gradio 
uvicorn